Reduce failure surface for gradle.
If the config files (build_apk.cmake or any of the android project
templates) change, this will force gradle to update its dependencies
from the internet. Otherwise gradle will always run in offline mode.

Furthermore this moves the gradle cache folder into the build directory,
so we do not pollute the global gradle build system with our files. Also
it means that we will not conflict with normal gradle builds if we
cannot get one of the file locks.

Change-Id: I9a8996d2fa824e8184aa81996af8707332b95fde
Reviewed-on: https://team-review.git.corp.google.com/50206
Reviewed-by: David Neto <dneto@google.com>