Update the function signature of {Fill|Dump}ImageLayersData
FillImageLayersData() now returns a tuple of bool and
vulkan::VkcommandBuffer that contains the recorded commands when the
operation can be done successfully. The command buffer should not be
freed before all the recorded commands get executed.

DumpImageLayersData() will create a local temporary command buffer
inside of it. It is valid to destroy the command buffer as this function
will wait for queue idle before it returns.

Change-Id: I2df45b4baf7fe63f25f079b2757e2b7176046a5f
Reviewed-on: https://team-review.git.corp.google.com/56003
Reviewed-by: Andrew Woloszyn <awoloszyn@google.com>